Sort the options in the tun config.
When creating a tun config with more than one global option using ruby 1.8 ( the default on ubuntu 12.04) it returns the hash in a random order. This means that the config file changes on most puppet runs and the service restarts.
This just sorts the options by name before looping though keeping them consistent.