Add support for multiple stunnel options
The options parameter previously defaulted to the empty string and any
value passed to it was added as a single 'options =' line. Support taking
a string for backwards compatability, but now default to an empty array
and accept an array of strings, each of which generates a separate
'options =' line.

Add supporting spec test and update the README.