Update README to fix info. on client cert usage
The README incorrectly stated that the certificate file on the client
would be used in the two computer example. The example also encouraged
people to install the server's private key on the client, which is not
good practice.

stunnel's default for `verify` is "No verify". Therefore, in the
example, the client would present the server's certificate as its own
but the server would ignore it anyway.

Now that certificates can be made optional for clients, it is clearer to
simply omit the client certificate from the example.